MVC Structure Prep
I’ve altered the structure of the code based on feedback from tutors.
The Model and View need to have interfaces and are to be created from
the controller which may or may not need an interface (probably won’t).

View should implement Observer and Model Observable and will probably
need multiple classes for each.

Removed the excess interfaceController class as it was causing
confusion and was redundant.

From this point onward Interface will be used in reference to the
programming context whereas GUI shall be used when referencing
graphical content