Moved components in GUI & added comments
Separated the add stock elements of the GUI from the total stock and
history fields. Still can’t get everything to display without a resize
which seems to be a common problem… Continued with commenting each
element of the GUI, includes some notes on previous layout positions in
case anything needs reverted and to help me keep track.