Added frames for History Window & EditStock Class
Made default frames so that we can work on building these other windows
without impacting on the main project code. Sorry for all the main
functions!

Also added action listener to main gui to call the edit stock window.
Not sure if this should be in a separate class or not. If it’s only
going to call the window it should be fine as it is, but if we need to
pass data it’ll need its own class.